mcp/requirements.txt
fastapi
uvicorn
requests
bunq-sdk
python-dotenv
python-picnic-api2
